Pkg restructure: removed refs to util_computer.py
