+++ aiida_jutools/util_code.py
@@ -19,7 +19,8 @@ import typing as _typing
 import aiida as _aiida
 import aiida.orm as _orm
 
-from aiida_jutools import computer as _jutools_computer
+import aiida_jutools as _jutools
+
 
 def get_code(computer_name_pattern: str = "",
              code_name_pattern: str = "",
@@ -121,7 +122,7 @@ def get_code(computer_name_pattern: str = "",
 
         return codestring, error_msg
 
-    computers = _jutools_computer.get_computers(computer_name_pattern)
+    computers = _jutools.computer.get_computers(computer_name_pattern)
     if not computers:
         raise _aiida.common.exceptions.NotExistent(f"No computer '{computer_name_pattern}' found.")
     else:
@@ -141,7 +142,7 @@ def get_code(computer_name_pattern: str = "",
 
         # first try to get queue_name from computer
         try:
-            queue_name = _jutools_computer.get_least_occupied_queue(computer=computer,
+            queue_name = _jutools.computer.get_least_occupied_queue(computer=computer,
                                                                     gpu=None,
                                                                     with_node_count=False,
                                                                     silent=True)
@@ -187,9 +188,9 @@ def get_code(computer_name_pattern: str = "",
                           f"Will choose first one. If this is a problem, contact developer.")
 
             queues_harcoded = list(computer_queue_architectures.keys())
-            queues_queried = _jutools_computer.get_queues(computer=computer,
-                                                          gpu=None,
-                                                          with_node_count=False)
+            queues_queried = _jutools.computer.get_queues(computer=computer,
+                                                      gpu=None,
+                                                      with_node_count=False)
             if not set(queues_harcoded) == set(queues_queried):
                 raise ValueError(
                     f"Computer '{computer_name_pattern}' hardcoded queues {queues_harcoded} do not "
@@ -202,9 +203,9 @@ def get_code(computer_name_pattern: str = "",
             if not architecture:
                 # since have just that hardcoded queues match actual queues, can conclude
                 # that user has specified non-existant queue
-                module_name = _inspect.getmodulename(_inspect.getfile(_jutools_computer.get_queues))
+                module_name = _inspect.getmodulename(_inspect.getfile(_jutools.computer.get_queues))
                 raise KeyError(f"Computer '{computer_name_pattern}' has no queue '{queue_name}'. Use "
-                               f"'{module_name}.{_jutools_computer.get_queues.__name__}()' to get list of queues.")
+                               f"'{module_name}.{_jutools.computer.get_queues.__name__}()' to get list of queues.")
 
             cs_by_computer_code_arch = [cs for cs in cs_by_computer_code if architecture in cs]
